feat: Add ft_select_cmd.c to executor module
The code changes include adding a new file `ft_select_cmd.c` to the `executor` module. This file contains the implementation of a function that is necessary for the execution of commands.

Based on the recent user commits and repository commits, it seems that this addition is part of ongoing development work on the project.